Create instanton.ld.so.cache in a separate stage
Creating an ld.so.cache that excludes processor-specific Glibc libraries
for InstantOn currently requires modifying the base image (by
temporarily renaming system directories, and thereby affecting
last-modified-time) because of limitations in ldconfig. This is
undesirable because we don't otherwise modify the base image and don't
want to unecessarily appear to.

This patch fixes the issue by breaking the build into multiple stages
and generating the cache in a separate stage that is then discarded,
thereby leaving the final image unaffected.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>